refactor: add reader module
